add "exit" Cargo feature
When this feature is enabled the panic handler performs an exit semihosting call
after logging the panic message. This is useful when emulating the program on
QEMU as it causes the QEMU process to exit with a non-zero exit code; thus it
can be used to implement Cortex-M tests that run on the host.